move the cross-org test fixtures into Pest.php
foreignServer()/foreignSnapshot() and foreignRestore() built the same graph in
two test files, declared as Pest globals inside those files so availability
depended on load order. Move them next to the other shared fixture helpers, and
derive the restore's target from the snapshot rather than rebuilding the server.
