+++ tests/Feature/Livewire/Restore/IndexTest.php
@@ -4,11 +4,9 @@ use App\Enums\Ability;
 use App\Livewire\Restore\Index;
 use App\Models\BackupJob;
 use App\Models\DatabaseServer;
-use App\Models\Organization;
 use App\Models\Restore;
 use App\Models\Snapshot;
 use App\Models\User;
-use App\Models\Volume;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Livewire\Livewire;
 
@@ -34,29 +32,6 @@ function makeRestore(array $attrs = []): Restore
         'snapshot_id' => $snapshot->id,
         'target_server_id' => $target->id,
         'schema_name' => $attrs['schema_name'] ?? 'restored_db',
-    ]);
-}
-
-/**
- * A completed restore owned by an organization the actor is not a member of.
- */
-function foreignRestore(): Restore
-{
-    $org = Organization::factory()->create();
-    $server = DatabaseServer::factory()->create([
-        'organization_id' => $org->id,
-        'database_type' => 'mysql',
-    ]);
-    $snapshot = Snapshot::factory()
-        ->forServer($server)
-        ->onVolumes(Volume::factory()->create(['organization_id' => $org->id]))
-        ->create();
-
-    return Restore::create([
-        'backup_job_id' => BackupJob::create(['status' => 'completed'])->id,
-        'snapshot_id' => $snapshot->id,
-        'target_server_id' => $server->id,
-        'schema_name' => 'victim_confidential_schema',
     ]);
 }
